Add functions to return perlin noise in numpy arrays.
Since the perlin noise functions have known dimensions, I decided to just use for loops to index into their respective arrays and skip the use of an iterator.
The functions add size and resolution - size for the dimensions of the output, and resolution to determine the distance travelled along an axis - eg. if I specify a size of 10 and a resolution of 10, each element will be 1 unit away from the last.